Issue #1404: moving some tests, improvement of assertion method
