+++ api/src/test/java/com/epam/pipeline/util/CustomAssertions.java
@@ -16,6 +16,8 @@
 
 package com.epam.pipeline.util;
 
+import lombok.SneakyThrows;
+
 import java.util.function.Predicate;
 
 import static org.junit.Assert.fail;
@@ -64,17 +66,11 @@ public final class CustomAssertions {
 
     public static void assertThrowsChecked(final Class<? extends Throwable> expectedExceptionClass,
                                            final CheckedRunnable runnable) {
-        try {
-            runnable.run();
-        } catch (Exception e) {
-            if (expectedExceptionClass.isInstance(e)) {
-                return;
-            } else {
-                throw new AssertionError(
-                        String.format("Expected exception %s was not thrown, but another exception was: %s.",
-                                expectedExceptionClass.getSimpleName(), e.getClass().getSimpleName()), e);
-            }
-        }
-        fail(String.format("Expected exception %s was not thrown.", expectedExceptionClass.getSimpleName()));
+        assertThrows(expectedExceptionClass, () -> sneakyThrowing(runnable));
+    }
+
+    @SneakyThrows
+    private static void sneakyThrowing(final CheckedRunnable runnable) {
+        runnable.run();
     }
 }
